### Improving render performance by 237ms!

Improving render performance by 237ms by using the following perf tactics:

Start: 541ms

Render time after making fs-alert load only if error is called: 488ms

Render time after making dialog content only render when opening: 316.68ms

Render time after using strip-whitespace on template tag: 303.8

Final: 303.8