@liechtym 2017-04-24
I am going to close this PR until we have a standard way to handle pulling components in after render.